add garbage collector, memfunc, number parser, word_position, pointer keeper

+++ src/ft_atod_ex.c
@@ -2,5 +2,5 @@
 
 double		ft_atod_ex(const char *number)
 {
-	return (ft_atold_ex(number));
+	return ((double)ft_atold_ex(number));
 }
